Added functionality to delete a node Extracted default return objects into separate file
